Finish waitpid

+++ Userland/SampleCodeModule/include/sys_calls.h
@@ -66,7 +66,7 @@ uint64_t sys_unblock(uint64_t pid);
 
 uint64_t sys_yield();
 
-uint64_t sys_sem_open(char* sem_name, int init_value);
+int sys_sem_open(char* sem_name, int init_value);
 
 uint64_t sys_sem_close(void *sem);
